Complete the migration from vanilla css to mui
