--- src/components/NavItem/NavItem.css
@@ -1,16 +1,0 @@
-.nav__link {
-  display: flex;
-  align-items: center;
-  margin-left: 5px;
-  padding-left: 5px;
-  cursor: pointer;
-}
-.link-text {
-  transition: transform 0.3s ease-out;
-}
-.show-nav-item-text {
-  transform: translateX(0);
-}
-.hide-nav-item-text {
-  transform: translateX(-300%);
-}
