--- src/components/NavList/NavList.css
@@ -1,5 +1,0 @@
-.nav__list {
-  display: flex;
-  flex-direction: column;
-  align-items: space-between;
-}
