+++ src/components/index.js
@@ -1,4 +1,4 @@
-export * from "./NavDrawer/NavDrawer";
-export * from "./Layout/Layout";
-export * from "./FormInputList/FormInputList";
-export * from "./DoctorInfo/DoctorInfo";
+export * from "./NavDrawer";
+export * from "./Layout";
+export * from "./FormInputList";
+export * from "./DoctorInfo";
