--- src/pages/Signin/Signin.css
@@ -1,4 +1,0 @@
-.text-small {
-  font-size: var(--text-small);
-  font-weight: 300;
-}
