+++ src/pages/Signup.jsx
@@ -1,28 +1,26 @@
-import React, { useState } from "react";
-import { FormInputList } from "../../components";
-import { signinSchema } from "../../utility/formSchemas";
 import axios from "axios";
-import {
-  Button,
-  Container,
-  Typography,
-  Checkbox,
-  Link,
-  Box,
-} from "@mui/material";
-import "./Signin.css";
+import React, { useState } from "react";
+import { signupSchema } from "../utility/formSchemas";
+import { FormInputList } from "../components";
+import { Container, Link, Typography, Button } from "@mui/material";
 
 const formInitialState = {
+  name: "",
+  email: "",
   username: "",
   password: "",
+  phoneNumber: "",
 };
 const errorsInitialState = {
+  name: "",
+  email: "",
   username: "",
   password: "",
+  phoneNumber: "",
 };
-function SignIn() {
-  const [errors, setErrors] = useState(errorsInitialState);
+function SignUp() {
   const [user, setUser] = useState(formInitialState);
+  const [errors, setErrors] = useState(errorsInitialState);
   const handleInputChange = (e) => {
     setUser({ ...user, [e.target.name]: e.target.value });
     setErrors({ ...errors, [e.target.name]: "" });
@@ -30,7 +28,7 @@ function SignIn() {
   const handleSubmit = async (e) => {
     e.preventDefault();
     try {
-      const { error } = signinSchema.validate(user, {
+      const { error } = signupSchema.validate(user, {
         abortEarly: false,
       });
       if (error) {
@@ -42,7 +40,7 @@ function SignIn() {
         return;
       }
       await axios.post(
-        `${import.meta.env.VITE_BASE_URL}user/patient/signin`,
+        `${import.meta.env.VITE_BASE_URL}user/patient/signup`,
         user
       );
     } catch (error) {
@@ -51,11 +49,25 @@ function SignIn() {
   };
   const formInputs = [
     {
+      placeholder: "Full Name",
+      value: user.name,
+      name: "name",
+      onChange: handleInputChange,
+      type: "text",
+    },
+    {
       placeholder: "Username",
       value: user.username,
       name: "username",
       onChange: handleInputChange,
       type: "text",
+    },
+    {
+      placeholder: "Email",
+      value: user.email,
+      name: "email",
+      onChange: handleInputChange,
+      type: "email",
     },
     {
       placeholder: "Password",
@@ -74,7 +86,7 @@ function SignIn() {
           sx={{ textAlign: "center" }}
           fontWeight="bold"
         >
-          Sign in
+          Sign up
         </Typography>
         <Typography
           variant="subtitle1"
@@ -82,28 +94,33 @@ function SignIn() {
           sx={{ textAlign: "center" }}
           fontWeight="bold"
         >
-          Log into your account
+          Create an account
         </Typography>
         <FormInputList
           formInputs={formInputs}
           errors={errors}
           changeHandler={handleInputChange}
         />
-        <Button variant="contained" color="primary" type="submit">
-          Sign in
+        <Button
+          variant="contained"
+          color="primary"
+          type="submit"
+          sx={{ width: "100%", my: 2 }}
+        >
+          Sign up
         </Button>
-        <Box sx={{ display: "flex", justifyContent: "space-between", my: 2 }}>
-          <Box>
-            <Checkbox sx={{ p: 0, mr: 1 }} />
-            <Typography variant="caption">Remember me</Typography>
-          </Box>
-          <Link href="/login" underline="always">
-            Don't have an account? Join us
-          </Link>
-        </Box>
+        <Link href="/signin">
+          <Typography
+            variant="subtitle1"
+            color="primary"
+            sx={{ textAlign: "center" }}
+          >
+            Have an account? Login
+          </Typography>
+        </Link>
       </form>
     </Container>
   );
 }
 
-export { SignIn };
+export { SignUp };
